+++ src/HttpClient/typings.ts
@@ -4,7 +4,7 @@ import { Span } from 'opentracing'
 
 import { CacheLayer } from '../caches/CacheLayer'
 import { MetricsAccumulator } from '../metrics/MetricsAccumulator'
-import { UserLandTracer } from '../tracing/UserLandTracer'
+import { IUserLandTracer } from '../tracing/UserLandTracer'
 import { Cached, CacheType } from './middlewares/cache'
 
 export type InflightKeyGenerator = (x: RequestConfig) => string
@@ -44,8 +44,8 @@ export interface RequestConfig extends AxiosRequestConfig {
 
 export interface TraceableRequestConfig extends RequestConfig {
   tracing?: {
-    rootSpan?: Span,
-    tracer: UserLandTracer
+    rootSpan?: Span
+    tracer: IUserLandTracer 
   }
 }