+++ src/index.ts
@@ -1,5 +1,7 @@
 process.env.FORCE_COLOR = '1'
 
+import * as tracingUtils from './tracing'
+
 export * from './caches'
 export * from './clients'
 export * from './errors'
@@ -16,3 +18,5 @@ export * from './service'
 export * from './service/logger'
 export * from './utils'
 export * from './constants'
+export { tracingUtils }
+